Split the treemap generator into focused modules
